# src/Studio/AterStudio/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AterStudio API doc",
    "description": "Studio API 文档. 更新时间:2024-11-03 17:21:10",
    "version": "v1"
  },
  "paths": {
...
    "/api/admin/ApiDocInfo/{id}": {
      "get": {
        "tags": [
          "ApiDocInfo"
        ],
        "summary": "获取某个文档信息",
        "operationId": "ApiDocInfo_GetApiDocContent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "isFresh",
            "in": "query",
            "description": "",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocContent"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ApiDocInfo"
        ],
        "summary": "更新",
        "operationId": "ApiDocInfo_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiDocInfoUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiDocInfoUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApiDocInfoUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ApiDocInfo"
        ],
        "summary": "删除",
        "operationId": "ApiDocInfo_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/ApiDocInfo/export/{id}": {
      "get": {
        "tags": [
          "ApiDocInfo"
        ],
        "summary": "导出markdown文档",
        "operationId": "ApiDocInfo_Export",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/ApiDocInfo/component": {
      "post": {
        "tags": [
          "ApiDocInfo"
        ],
        "summary": "生成页面组件",
        "operationId": "ApiDocInfo_CreateUIComponent",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUIComponentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUIComponentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUIComponentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NgComponentInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/ApiDocInfo/generateRequest/{id}": {
      "get": {
        "tags": [
          "ApiDocInfo"
        ],
        "summary": "生成前端请求",
        "operationId": "ApiDocInfo_GenerateRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "webPath",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "schema": {
              "$ref": "#/components/schemas/RequestLibType"
            }
          },
          {
            "name": "swaggerPath",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/GenAction/filter": {
      "post": {
        "tags": [
          "GenAction"
        ],
        "summary": "分页数据",
        "operationId": "GenAction_Filter",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenActionFilterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenActionFilterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenActionFilterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenActionItemDtoPageList"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/GenAction/steps/{id}": {
      "get": {
        "tags": [
          "GenAction"
        ],
        "summary": "获取操作步骤",
        "operationId": "GenAction_GetSteps",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GenStepItemDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "GenAction"
        ],
        "summary": "添加操作步骤",
        "operationId": "GenAction_AddSteps",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/GenAction/modelFile": {
      "get": {
        "tags": [
          "GenAction"
        ],
        "summary": "获取模型列表",
        "operationId": "GenAction_GetModelFile",
        "parameters": [
          {
            "name": "sourceType",
            "in": "query",
            "description": "",
            "schema": {
              "$ref": "#/components/schemas/GenSourceType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelFileItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/GenAction/{id}": {
      "patch": {
        "tags": [
          "GenAction"
        ],
        "summary": "更新数据",
        "operationId": "GenAction_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenActionUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenActionUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenActionUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "GenAction"
        ],
        "summary": "获取详情",
        "operationId": "GenAction_GetDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenActionDetailDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "GenAction"
        ],
        "summary": "删除",
        "operationId": "GenAction_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/GenAction/execute": {
      "post": {
        "tags": [
          "GenAction"
        ],
        "summary": "执行操作",
        "operationId": "GenAction_Execute",
        "requestBody": {
          "description": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenActionRunDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenActionRunDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenActionRunDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "NotStarted",
          "InProgress",
          "Success",
          "Failed"
        ],
        "x-enumData": [
          {
            "name": "NotStarted",
            "value": 0,
            "description": "未执行"
          },
          {
            "name": "InProgress",
            "value": 1,
            "description": "执行中"
          },
          {
            "name": "Success",
            "value": 2,
            "description": "成功"
          },
          {
            "name": "Failed",
            "value": 3,
            "description": "失败"
          }
        ]
      },
      "ApiDocContent": {
        "type": "object",
        "properties": {
          "restApiGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RestApiGroup"
            }
          },
          "modelInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            }
          },
          "openApiTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiDocTag"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ApiDocInfoUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "文档名称",
            "nullable": true
          },
          "description": {
            "maxLength": 1000,
            "type": "string",
            "description": "文档描述",
            "nullable": true
          },
          "path": {
            "maxLength": 300,
            "type": "string",
            "description": "文档地址",
            "nullable": true
          },
          "localPath": {
            "maxLength": 200,
            "type": "string",
            "description": "生成路径",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "接口文档更新时请求结构"
      },
      "ApiDocTag": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ComponentType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Form",
          "Table",
          "Detail"
        ],
        "x-enumData": [
          {
            "name": "Form",
            "value": 0,
            "description": "提交表单"
          },
          {
            "name": "Table",
            "value": 1,
            "description": "展示表格"
          },
          {
            "name": "Detail",
            "value": 2,
            "description": "详情字段"
          }
        ]
      },
...
      "CreateUIComponentDto": {
        "required": [
          "modelInfo",
          "serviceName"
        ],
        "type": "object",
        "properties": {
          "uiType": {
            "$ref": "#/components/schemas/UIType"
          },
          "componentType": {
            "$ref": "#/components/schemas/ComponentType"
          },
          "modelInfo": {
            "$ref": "#/components/schemas/ModelInfo"
          },
          "serviceName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "GenActionDetailDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 40,
            "type": "string",
            "description": "action name"
          },
          "description": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "entityPath": {
            "type": "string",
            "description": "实体路径",
            "nullable": true
          },
          "openApiPath": {
            "type": "string",
            "description": "open api path",
            "nullable": true
          },
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Variable"
            }
          },
          "sourceType": {
            "$ref": "#/components/schemas/GenSourceType"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "actionStatus": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "生成操作详情"
      },
      "GenActionFilterDto": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            },
            "nullable": true
          },
          "name": {
            "maxLength": 40,
            "type": "string",
            "description": "action name",
            "nullable": true
          },
          "sourceType": {
            "$ref": "#/components/schemas/GenSourceType"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "actionStatus": {
            "$ref": "#/components/schemas/ActionStatus"
          }
        },
        "additionalProperties": false,
        "description": "生成操作筛选条件"
      },
...
      "GenActionItemDtoPageList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenActionItemDto"
            }
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GenActionRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "操作id",
            "format": "uuid"
          },
          "sourceFilePath": {
            "type": "string",
            "description": "源文件路径",
            "nullable": true
          },
          "isFile": {
            "type": "boolean",
            "description": "是否写入文件"
          }
        },
        "additionalProperties": false,
        "description": "操作执行模型"
      },
      "GenActionUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 40,
            "type": "string",
            "description": "action name",
            "nullable": true
          },
          "description": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "entityPath": {
            "type": "string",
            "description": "实体路径",
            "nullable": true
          },
          "openApiPath": {
            "type": "string",
            "description": "open api path",
            "nullable": true
          },
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Variable"
            },
            "nullable": true
          },
          "sourceType": {
            "$ref": "#/components/schemas/GenSourceType"
          },
          "actionStatus": {
            "$ref": "#/components/schemas/ActionStatus"
          }
        },
        "additionalProperties": false,
        "description": "生成操作更新时DTO"
      },
      "GenSourceType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "EntityCLass",
          "DtoModel",
          "OpenAPI"
        ],
        "x-enumData": [
          {
            "name": "EntityCLass",
            "value": 0,
            "description": "实体类"
          },
          {
            "name": "DtoModel",
            "value": 1,
            "description": "Dto模型"
          },
          {
            "name": "OpenAPI",
            "value": 2,
            "description": "OpenAPI"
          }
        ]
      },
...
      "GenStepItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "步骤名称"
          },
          "outputPath": {
            "maxLength": 400,
            "type": "string",
            "description": "输出路径",
            "nullable": true
          },
          "path": {
            "maxLength": 400,
            "type": "string",
            "description": "路径",
            "nullable": true
          },
          "genStepType": {
            "$ref": "#/components/schemas/GenStepType"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "task step列表元素"
      },
...
      "ModelFileItemDto": {
        "required": [
          "fullName",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "名称"
          },
          "fullName": {
            "type": "string",
            "description": "路径"
          }
        },
        "additionalProperties": false,
        "description": "模型文件项"
      },
      "ModelInfo": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "类名"
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "类注释",
            "nullable": true
          },
          "summary": {
            "maxLength": 100,
            "type": "string",
            "description": "类注释",
            "nullable": true
          },
          "keyType": {
            "$ref": "#/components/schemas/EntityKeyType"
          },
          "isEnum": {
            "type": "boolean",
            "description": "是否为枚举类",
            "nullable": true
          },
          "isList": {
            "type": "boolean"
          },
          "propertyInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyInfo"
            },
            "description": "属性"
          }
        },
        "additionalProperties": false,
        "description": "模型信息"
      },
...
      "NgComponentInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "routing": {
            "type": "string",
            "nullable": true
          },
          "tsContent": {
            "type": "string",
            "nullable": true
          },
          "htmlContent": {
            "type": "string",
            "nullable": true
          },
          "cssContent": {
            "type": "string",
            "nullable": true
          },
          "moduleName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RequestLibType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "NgHttp",
          "Axios"
        ],
        "x-enumData": [
          {
            "name": "NgHttp",
            "value": 0,
            "description": "angular http"
          },
          {
            "name": "Axios",
            "value": 1,
            "description": "axios"
          }
        ]
      },
      "RestApiGroup": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "apiInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RestApiInfo"
            }
          }
        },
        "additionalProperties": false,
        "description": "接口分组信息"
      },
...
      "UIType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "AngularMaterial",
          "NGZORRO",
          "VueElement"
        ],
        "x-enumData": [ ]
      },
...
      "Variable": {
        "required": [
          "key",
          "value"
        ],
        "type": "object",
        "properties": {
          "key": {
            "maxLength": 100,
            "type": "string"
          },
          "value": {
            "maxLength": 1000,
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}